+++ ClientApp/src/components/wishList/buttons/AddNewWishButton.tsx
@@ -1,4 +1,4 @@
-import styles from './WishItem.module.scss';
+import styles from './WishButton.module.scss';
 import { useHistory } from "react-router";
 
 const AddNewWishButton = () => {
@@ -10,7 +10,7 @@ const AddNewWishButton = () => {
 
   return (
     <div className={styles.addNewWishButton}>
-      <button onClick={handleClick}>Add New Wish</button>
+      <button onClick={handleClick} className={styles.lightButton}>Add New Wish</button>
     </div>
   );
 }
