+++ ClientApp/src/components/wishList/buttons/WishButton.module.scss
@@ -1,0 +1,81 @@
+$wish-color: #D69CCF;
+
+/*light button*/
+
+.lightButton {
+	background-color:#ffffff;
+	display:inline-block;
+	cursor:pointer;
+	color:$wish-color;
+	font-family:Arial;
+	font-size:15px;
+	padding:9px 23px;
+	text-decoration:none;
+	text-shadow:0px -1px 0px $wish-color;
+  border: 0;
+
+  box-shadow: 0px 0px 7px 0px $wish-color;
+}
+
+.lightButton:hover {
+	background:linear-gradient(to bottom, #fffaff 5%, #edc2ed 100%);
+	background-color:#fffaff;
+}
+
+.lightButton:active {
+	position:relative;
+	top:1px;
+}
+
+/*contrast button*/
+
+.contrastButton {
+  box-shadow:inset 0px 34px 0px -15px #e0b6db;
+	background:linear-gradient(to bottom, #ffffff 5%, #d69ccf 100%);
+	background-color:#ffffff;
+	border: 0;
+	display:inline-block;
+	cursor:pointer;
+	color:#ffffff;
+	font-family:Arial;
+	font-size:15px;
+	padding: 5px 25px;
+	text-decoration:none;
+	text-shadow:0px -1px 0px #c763bd;
+}
+
+.contrastButton:hover {
+	background:linear-gradient(to bottom, $wish-color 5%, #ffffff 100%);
+	background-color:$wish-color;
+}
+
+.contrastButton:active {
+	position:relative;
+	top:1px;
+}
+
+/* view button */
+
+.viewButton {
+  margin-top: 35px;
+}
+
+/* add new wish button */
+
+.addNewWishButton {
+  margin: 35px 10px;
+}
+
+/* add new wish button */
+
+.submitButton {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-template-rows: 1fr;
+  grid-column-gap: 0px;
+  grid-row-gap: 0px;
+}
+
+.submitButton button {
+  grid-area: 1 / 2 / 2 / 3;
+}
